LayoutTests/editing/pasteboard/paste-global-selection-expected.txt


This test checks that we can set and paste X11-style selections.

PASS confirmedMarkup is 'TestTest'
PASS confirmedMarkup is 'SayHello'
PASS successfullyParsed is true

TEST COMPLETE
